BackEnd/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["yaseen1998 <yasseenkouthe@eng.hu.edu.jo>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.0.0"
django-cors-headers = "^3.13.0"
psycopg2-binary = "^2.9.4"

[tool.poetry.dev-dependencies]
Django = "^4.1.2"
black = "^22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
